Refactoring dashboard (#3)
* refactoring

* add lines to gitignore

* fix build error

* refactoring

* refactoring dashboard
