+++ src/components/end-user/achievement/components/MobileVerticalLineBottom.tsx
@@ -1,8 +1,8 @@
 // UI
-import { Box, useTheme } from '@chakra-ui/react'
+import { Box } from '@chakra-ui/react'
 
-// Utils
-import { getColor } from '@components/end-user/achievement/helpers'
+// Hooks
+import useGetColor from '../useGetColor'
 
 // Types
 import { IDesktopVerticalLine } from './DesktopVerticalLine'
@@ -11,8 +11,9 @@ interface IMobileVerticalLineBottom extends IDesktopVerticalLine {}
 
 const MobileVerticalLineBottom = (props: IMobileVerticalLineBottom) => {
   const { index, achievementsArray, achievement } = props
-const theme = useTheme()
-const bgColor = getColor(theme, achievement)
+
+  const { getColor } = useGetColor()
+
   return (
     <Box
       className="middle-vertical-line-below-circle"
@@ -22,7 +23,7 @@ const bgColor = getColor(theme, achievement)
       zIndex="-1"
     >
       <Box
-        bg={index === achievementsArray.length - 1 ?  'brand.neutral3' : bgColor} //getColor(achievement)
+        bg={index === achievementsArray.length - 1 ? 'brand.neutral3' : getColor(achievement)}
         h="120px"
         position="absolute"
         w="1px"
