+++ src/components/end-user/achievement/components/MobileVerticalTop.tsx
@@ -1,8 +1,8 @@
 // UI
-import { Box, useTheme } from '@chakra-ui/react'
+import { Box } from '@chakra-ui/react'
 
-// Utils
-import { getColor } from '@components/end-user/achievement/helpers'
+// Hooks
+import useGetColor from '../useGetColor'
 
 // Types
 import { IDesktopVerticalLine } from './DesktopVerticalLine'
@@ -13,25 +13,26 @@ import { achievementsArray } from '../constants'
 interface IMobileVerticalTop extends Omit<IDesktopVerticalLine, 'achievementsArray'> {}
 
 const MobileVerticalTop = ({ index, achievement }: IMobileVerticalTop) => {
-  const theme = useTheme()
-const bgColor = getColor(theme, achievement)
+  const { getColor } = useGetColor()
+
   return (
-  <Box
-    alignItems="center"
-    className="mobile-middle-vertical-line-top"
-    display="flex"
-    justifyContent="center"
-    position="relative"
-    w="100%"
-    zIndex="-1"
-  >
     <Box
-      bg={index === achievementsArray.length - 1 ? 'brand.neutral3' : bgColor}
-      h="160px"
-      position="absolute"
-      w="1px"
-    />
-  </Box>
-)}
+      alignItems="center"
+      className="mobile-middle-vertical-line-top"
+      display="flex"
+      justifyContent="center"
+      position="relative"
+      w="100%"
+      zIndex="-1"
+    >
+      <Box
+        bg={index === achievementsArray.length - 1 ? 'brand.neutral3' : getColor(achievement)}
+        h="160px"
+        position="absolute"
+        w="1px"
+      />
+    </Box>
+  )
+}
 
 export default MobileVerticalTop
