+++ src/components/end-user/dashboard/UserTierLevel/usePrevious.ts
@@ -1,0 +1,14 @@
+// Modules
+import { useEffect, useRef } from 'react'
+
+const usePrevious = (value) => {
+  const ref = useRef()
+
+  useEffect(() => {
+    ref.current = value
+  }, [value])
+
+  return ref.current
+}
+
+export default usePrevious
